--- src/components/task.d.ts
@@ -1,1 +1,0 @@
-declare module 'Task';
