+++ src/models/ITask.ts
@@ -1,0 +1,10 @@
+interface ITask {
+    readonly id: number,
+    title: string,
+    completed: boolean,
+    important: boolean,
+    deleted: boolean,
+}
+
+export default ITask;
+
