+++ src/models/TTask.ts
@@ -1,0 +1,10 @@
+type TTask = {
+    readonly id: number,
+    title: string,
+    completed: boolean,
+    important: boolean,
+    deleted: boolean,
+};
+
+export default TTask;
+
